feat(design) - убрал моки, поменял модели, внедрил IHttpClientFactory

+++ api/src/Sibintek.BeerMachine/Domain/Wallet.cs
@@ -2,7 +2,7 @@ namespace Sibintek.BeerMachine.Domain
 {
     public class Wallet
     {
-        public decimal Balance { get;  set; }
+        public long Balance { get;  set; }
         
         public long Id { get; set; }
     }
